requirements.txt
Django==4.2.7
pymongo
python-decouple==3.8
dnspython
